requirements.txt
requests==2.31.0
python-dotenv==1.0.0
mysql-connector-python==8.2.0
flask==3.0.0
flask-session==0.5.0
pandas==2.1.0